requirements.txt
streamlit==1.27.0
openai==1.3.0
youtube-transcript-api==0.6.1
python-dotenv==1.0.0